backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
pandas
python-dotenv
pydantic
tenacity
google-genai
supabase
huggingface_hub>=0.28.0
scikit-learn
